Divert Start/Continue link for Overseas Passports
Add logic to send users to hardcoded link when using the Overseas
Passports Smart Answer. This is a temporary measure while we work to
retire the Smart Answer permanently. We’ve had to roll out this
workaround as a matter of urgency.

There are no tests for the helper, which follows on from previous
changes on the ApplicationHelper in:

https://github.com/alphagov/smart-answers/pull/2792/commits/272251a8d302dd434c062385aa5ccfed81f0a075