Publish transaction edition via publishing api
In light of recent retire smartanswer requests (i.e overseas-passports):

- To reserve the base path of the smartanswer, after it has been
unpublished.
- Retain the start page structure, look and feel
- Link the start button to an external link

After testing a few mainstream publisher formats, the transaction format
appeared the mostly suitable to match the aforementioned points.

This commit contains a rake task that reserves and publishes a transaction.

This uses the publishing-api to first, change the publishing application
and publish a transaction in place of the smartanswer.

(i.e `rake retire:publish_transaction[base_path,publishing_app,title,content,link]  #`)

The rake task creates a transaction format, this uses the publishing api
to publish this the transaction.

When this rake task is invoked requests (i.e [1] and [2] from [3]) are
composed and made. This request publishes the transaction with the
supplied parameters.

For an example of a valid transaction format payload, see [4].

Tests have been added that describe this implementation.

[1] https://github.com/alphagov/publishing-api/blob/master/doc/api.md#put-pathsbase_path
[2] https://github.com/alphagov/publishing-api/blob/master/doc/api.md#put-v2contentcontent_id
[3] https://github.com/alphagov/publishing-api/blob/master/doc/api.md#post-v2contentcontent_idpublish
[4] https://github.com/alphagov/govuk-content-schemas/blob/4285c8d1efc95d9db932b546e092139e9a6022ae/formats/transaction/publisher_v2/examples/transaction.json