Publish answer edition via publishing api
Based on requests to retire overseas-passports smart answer with the
following required

- To reserve the base path of the smartanswer, after it has been
unpublished.
- Retain the start page structure, look and feel
- Link the start button to an external link

This commit contains a rake task that reserves and publishes an answer.

This uses the publishing-api to first, change the publishing application
and publish an answer in place of the smartanswer.

(i.e rake retire:publish_answer[base_path,publishing_app,title,content])

The rake task creates an answer content item, this uses the publishing api
to publish this.

When this rake task is invoked requests (i.e [1] and [2] from [3]) are
composed and made. This request publishes the start page in the answer
format with the supplied parameters.

For an example of a valid answer format payload, see [4].

Tests have been added that describe this implementation.

[1] https://github.com/alphagov/publishing-api/blob/master/doc/api.md#put-pathsbase_path
[2] https://github.com/alphagov/publishing-api/blob/master/doc/api.md#put-v2contentcontent_id
[3] https://github.com/alphagov/publishing-api/blob/master/doc/api.md#post-v2contentcontent_idpublish
[4] https://github.com/alphagov/govuk-content-schemas/blob/4285c8d1efc95d9db932b546e092139e9a6022ae/formats/answer/publisher_v2/examples/answer.json