Sort the files relating to a smart answer
This means that the return value of this method is no longer dependant
on the ordering on the underlying file system. Therefore, this helps
any functionality using all_paths or anything derived from it be more
deterministic.

The motivation for this change is the checksum generation output, as
with this change, the order of the files in the generated yaml file is
deterministic.