Implement checksums file management as two rake tasks
The generate-checksums-for-smart-answer script is used to create and
update the checksum files.

With the aim of making it possible to regenerate all the checksum
files, this commit refactors checksum file management as two rake
tasks. One to update the checksums for a flow, and another to create
or amend the checksum file itself.

This supports globbing, as the documentation for using the
generate-checksums-for-smart-answer script specified that it supported
adding directories (although I tried this, and it didn't work).