Add an AB test for calculate-your-child-maintenance
This sets up a test where we present alternate content to users in the B
variant.

We do this by configuring `ab_body` content for the pages that are affected.
At this point, we're only testing one page, but we expect to expand this to
encompass other content.

This test specifically removes an inline link from the "Before you start"
section, as we have noted that this distracts users from their ultimate goal,
which is probably starting the calculator.

This test is intended to run for between 1 and 2 weeks.

The corresponding Fastly configuration for this test was committed in
[this pull request](https://github.com/alphagov/fastly-configure/pull/21/files)

The ABTest-BenchmarkInlineLink cookie is listed on https://www.gov.uk/help/cookies