Fix calculate statutory sick pay tests
We have found that 2 tests for selecting the first sick day and last sick day
were failing as they were expecting an error to be raised if the date selected
is after the end of the current year.  However as of 1st June 2017 users are
able to select dates for the following year (see
https://github.com/alphagov/smart-answers/pull/2878), hence the tests are now
failing.  This fixes these tests to freeze the dates so that this issue
shouldn't occur again in future.