Handle errors from Imminence
When imminence doesn't respond with a status of ok, or when the api
adapter raises any error, report the error using Airbrake, and show a
general failure message to the end user.

Previously the failure modes would either be a server error from smart
answers, in the case that the request to imminence failed, or
returning false from rules_apply? in the case where imminence returned
a status of 404 in the _response_info.status.