Add an option for tenants from else where
This commit updates the `what_nationality?` question to allow users
to select an  option that identifies the tenant is from else where.

i.e "somewhere-else": "From somewhere else or you don’t know"

If the user selects the aforementioned option, they are sent to
`outcome_can_not_rent` outcome. This is used as a temporary stub
and will be changed in the next commit.

Unit tests have been updated to accept the `somewhere-else` option.
Regression tests have been updated to test this new option.