Create an outcome per path for Cyprus in Marriage Abroad
Copy the test artefacts in to the templates directory, rename the
files from .txt to .erb, remove the titles, and add the title
partial. Add cyprus to the list of countries with an outcome per path.

Update the checksums file, adding the new templates.

The test artefacts have not been changed.