Create an outcome per path for Poland
The goal of this story is to migrate the flow for Poland to use
the outcome per path strategy.

Within, flat content have been introduced for Poland. The flat
content will be linked and form part of the outcome per path flow.

Given the aforementioned, the test artefacts haven't been affected by
this change.

For opposite sex marriage in poland the addresses have not been pulled
from whitehall. This is following the example from [1]

[1] outcome_opposite_sex_marriage_in_poland_when_residing_in_uk