Test different button placement
We want to test the behaviour of users when the start button is higher up the
page.  We observed that on the current page, the button is below the fold and
also below several links that are tantalising for users, causing them to miss
the service.

This test will run until 23rd June 2017.