Amend disclaimer text
In user research we discovered that people wanted to know more about why it was
an estimate.  We think that this text is slightly clearer.

The disclaimer text now matches what is in the B variant for the CmButton1 test

This is a big commit because the disclaimer text is used in a lot of places, so
a number of the test artefacts have changed.

The partial has no newline to finish as this causes a problem with the
generated markup in a call out.