Add annual leave entitlement AB test
Our user research suggests that giving the context of when a leave year starts
closer to the place where a user needs that information could help completion.

The text we are testing is this:
```

  The leave year will usually be in an employee’s contract. If it isn’t, their leave year will start:

  - on their first day (if the job began after 1 October 1998)
  - on 1 October (if the job began on or before 1 October 1998)
```

The A variant keeps this text on the landing page where it currently.

The B variant does not show this text on the landing page, but instead on:

- https://www.gov.uk/calculate-your-holiday-entitlement/y/days-worked-per-week
- https://www.gov.uk/calculate-your-holiday-entitlement/y/shift-worker
- https://www.gov.uk/calculate-your-holiday-entitlement/y/hours-worked-per-week

To minimise duplication and ot make it more obvious where the text is displayed,
I've put this markup directly in the question view rather than using Govspeak
in the node body (which is the approach we've typically used for the landing
pages)

This test will run until at least 26th June 2017.

https://trello.com/c/GgnYfigi/161-annual-leave-description-develop-and-deploy-change-to-form-flow