Derive outcome per path countries
Prior to this, the countries_with_18_outcomes list held all the outcome
per path or countries that have been flattened.

This has led to duplication of countries with the other lists.

This commit changes this by adding implementation that derives outcome
per path countries from the existing lists.

This removes the duplication and hence presences a clean and more accurate
list of flattened countries.

The regression tests artefacts and others are not affected by the changes
in this commit