Refactor ContentItemPublisher
This ensures each time the #publish method is being used (which it is upon
deploy), the content-item living at "/part-year-profit-tax-credits" gets
unpublished, its draft gets deleted, and the start page of that smart-answer
gets published as a transaction. The rest of the smart-answer gets published as
a StartPageLessFlowContentItem under "/part-year-profit-tax-credits/y".