Copy outcomes and questions to part-year-profit-tax-credits/y
The controller expects to find the content for outcomes and questions under
"lib/smart_answer_flows/base-path" so we've had to copy the content into a
nested y/ folder for the views to work. This can be cleaned up once all smart
answer start pages have been moved to Publisher.