Update Child Maintenance page after AB testing
We're updating the text on the landing page and the disclaimer to reflect the
things we've found during this mission.

We're simplifying the text slightly and moving the start button to be above the
fold in more situations.

https://trello.com/c/lrUPjkFu/191-publish-deploy-final-child-maintenance-changes

The changes were agreed with the department as per this card: https://trello.com/c/TzvnQWif/187-agree-final-child-maintenance-changes-with-dept