Move custom start button text to ERB templates
And make them available through the `StartNodePresenter`. This is in
preparation for being able to send the custom Start Button text to the
Publishing API along with the rest of the content of the Start Page.

I ran the regression tests for calculate-your-child-maintenance and
confirmed that there were no changes to the rendered artefacts (i.e. the
custom start button text is still being used). As such I'm happy to
update the checksums for this Smart Answer. I also ran the regression
tests for the student-finance-forms Smart Answer to ensure that Start
Pages without custom start button text were still being rendered
correctly.

I've removed the StartButton and associated tests now that it's
redundant.

I've removed integration/start_page_test as it was relying on
`StartButton` and it's not clear to me that it serves any purpose now
that `StartButton` has been removed.