Remove related links' overrides
Recently, some transaction pages have been moved from smart-answers
into frontend. Two of those pages had overrides to show the old
related links instead of the new taxonomy related links when presented
in the B variant of the Education Navigation A/B test.

This means the overrides are no longer needed in this repository. This
commit removes the code related to this functionality.

There is more context in:
- https://trello.com/c/J8ofEnrt/167-display-curated-related-links-in-the-sidebar
- https://docs.google.com/document/d/1k1nlQOpFwjaUNoNtizAnsDr5z9O5uUMIOlgDSUCbwj4/edit#