Create an outcome per path for Azerbaijan
The goal of this story is to migrate the flow for Azerbaijan to
use the outcome per path strategy.

Within, flat content have been introduced for Azerbaijan. The flat content
will be linked and form part of the outcome per path flow.

Given the aforementioned, the test artefacts haven't been affected by
this change