Add docs about Heroku Review apps
For Factcheck we no longer need to use the startup_heroku.sh script
to deploy a preview of Smart Answers for each PR we raise, as we
have cofigured Heroku Review apps for the repo.

We have retained the manual script as a backup in case the automatic
process should go wrong for any reason. It is important to be able
to easily create the previews as we send these to government
departments to review.