Retrieve ContentItem via ContentRetriever Service
Sequel to the issue raised in [1], it appeared that recent AB testing
efforts have introduced code that makes it hard to create a new smart
answer. Existing smart answers have not been affected.

If the `smart-answer` is not in the content store, then content_item
instance variable is set to nil instead of the response from the
content store.

The idea here is to return a hash after the request, such that no matter
how many times it is exccuted, even if the slug/smart answer isn't
defined on the content store the same type (i.e Hash) is returned.

In theory, the aforementioned is also regarded as idempotence.

This commit moves the request for the content item to a class level
method called fetch defined within a ContentItemRetriever Service.

Also inline with the existing code, a without_links_organisations class
level method has also been defined to cater for the removal of
organisations under the links key.

Tests have also been written to cover the aforementioned areas,
including controller tests that ratify the assignment of
navigation_helper and content_item instance variables.

[1] https://github.com/alphagov/smart-answers/issues/3136