Create an outcome per path for El Salvador
The goal of this commit is to migrate the flow for El Salvador to use
the outcome per path strategy.

Within, flat content have been introduced for El Salvador. The flat
content will be linked and form part of the outcome per path flow.

Given the aforementioned, the test artefacts haven't been affected by
this change.