Templates for British Indian Ocean Territory
The goal of this commit is to migrate the flow for British Indian
Ocean Territory to use the outcome per path strategy.

Within, flat content have been introduced for British Indian Ocean
Territory. The flat content will be linked and form part of the outcome
per path flow.

Given the aforementioned, the test artefacts haven't been affected by
this change.