Preparatory: Generate test artefacts for Algeria
Given the need to reduce the complexity around editing marriage abroad
templates. Algeria has been picked to be converted to an outcome
per path country.

Prior to this, Algeria did not have any defined test artefacts.

This commit is a preparatory one, that generates regression test
artefacts for marriages in Algeria.

This will help track the changes from later commits.