Question 37: Replace Yes/No with dispute criteria
As part of the accessibility improvements, we have been ask to replace
the Yes/No in question 37 (i.e is_any_of_the_following_apply?) with the
dispute criteria listed above it. The idea is to assist devices like
screen reader etc, to better recognise the dispute criteria.

When a user selects at least on of the criteria they would be sent to
the "is not eligible" outcome. When none is selected and the next
steps button is triggered the user is shown "is abroad for treatment"
outcome.

This commit contains implementation and tests that achieves & describe
the aforementioned scenarios.

The  regression test artefacts have been affected by these changes
they will be added in the next commit.