Question 33: Replace Yes/No with partner premium
As part of the accessibility improvements, we have been ask to replace
the Yes/No in question 33 (i.e is_claiming_benefits?) with the
partner premium listed above it. The idea is to assist devices like
screen reader etc, to better recognise the partner premiums.

When a user selects at least on of the premiums they would be sent to
the "is claiming benefits" outcome and when none are select and the
next steps button is triggered the user is shown the is either of the
following question.

This commit contains implementation and tests that achieves & describe
the aforementioned scenario.

Regression test artefacts have been affected by these changes they will
be added in the next commit.

NB:

The option value in the erb template contain HTML anchor tags, this was
has been added to give users access to content that further explain
more about the premium in question. The content design team have asked
that these links be retained.