Question 34: Replace Yes/No with impairments
As part of the accessibility improvements, we have been ask to replace
the Yes/No in question 34 (i.e is_either_of_the_following?) with the
possible impairments listed above it. The idea is to assist devices
like screen reader etc, to better recognise the possible impairments.

When a user selects at least an impairment they would be sent to
the "is abroad for treatment" question and when none are select and
the next steps button is triggered the user is shown the "is any of
the following apply" question.

This commit contains implementation and tests that achieves & describe
the aforementioned scenario.

Regression test artefacts have been affected by these changes they will
be added in the next commit.