Question 36: Replace Yes/No with impairment periods
As part of the accessibility improvements, we have been ask to replace
the Yes/No in question 36 (i.e is_work_or_sick_pay?) with the periods
during which the said impairment had lasted. These hare already listed
above the options.

The idea here, is to assist devices like screen reader etc, to better
recognise the impairment periods.

When a user selects at least an impairment period they would be sent
to the "is_abroad_for_treatment" outcome and when none are select and
the next steps button is triggered the user is shown the "is not
eligible" outcome.

This commit contains implementation and tests that achieves & describe
the aforementioned scenario.

Regression test artefacts have been affected by these changes they will
be added in the next commit.