Question 20: Replace Yes/No with tax credits
As part of the accessibility improvements, we have been ask to replace
the Yes/No in question 20 (i.e tax_credits_currently_claiming?) with
tax credit benefits. These are already listed above the options.

The idea here, is to assist devices like screen reader etc, to better
recognise the tax credit benefits.

When a user selects at least an tax credit benefits they would be sent
to the "tax_credits_eea_entitled" outcome and when none are select and
the next steps button is triggered the user is shown the "tax credits
unlikely" outcome.

This commit contains implementation and tests that achieves & describe
the aforementioned scenario.

Regression test artefacts have been affected by these changes they will
be added in the next commit.