Question 2: Replace Yes/No with list of benefits
As part of the accessibility improvements, we have been ask to replace
the Yes/No in question 2 (i.e gets_benefits?) with the list of state
benefits listed above it. The idea is to assist devices like screen
reader etc, to better recognise the benefits.

When a user selects at least one of the benefits they would be sent to
the "how_many_nights_children_stay_with_payee?" question. When none is
selected and the next steps button is triggered the user is shown the
"gross_income_of_payee?" question.

This commit contains implementation and tests that archives & describe
the aforementioned.

Regression test artefacts have been affected by these changes they will
be added in the next commit.