Question 3: Replace Yes/No with exempt benefits
As part of the accessibility improvements, we have been ask to replace
the Yes/No in question 3 (i.e receiving_exemption_benefits?) with
exempt benefits. These are already listed above the options.

The idea here, is to assist devices like screen reader etc, to better
recognise the exempt benefits.

When a user selects at least an exempt benefits they would be sent
to the "not_affected_exemptions" outcome and when none are select and
the next steps button is triggered the user is shown the "receiving non
exemption benefits" question.

This commit contains implementation and tests that achieves & describe
the aforementioned scenario.

Regression test artefacts have been affected by these changes they will
be added in the next commit.