DRYUP: Update state benefits
This commit updates the state benefits list. It defines it as a
constant on the calculator and reuses it in the flow and template.

For the flow the calculator variable scope has been shared with the
state object and questions in the flow.

This makes it easy for the state benefits to be defined as the option
in question 13.