Raise an error if the Worldwide API does not return any countries
We recently had an incident where the Worldwide API was not returning
any countries to Smart Answers.

As such, the country select questions presented users with empty
dropdowns. This case is irrecoverable by the user, so we should fail
loudly.

This change checks if we receive any locations from the API. If not,
then we raise an exception. This will achieve two things:

- The error will be logged to Errbit
- The user will be presented with a generic 500 error page, informing
  them that something has gone wrong at our end.