Define and apply list validator.
Following the accessibility improvements, a few of repetitive predicate
methods have been defined and apply across the following smart answers:

- PR 3181 Benefit Cap Calculator
- PR 3180 Calculate your child maintenance

This commit defines a list validator that reduces the duplication
introduced during the accessibility improvements.

The idea here is to check the supplied argument and its contents exist
in the list that serves as a source of truth. Also that each element's
exists and its size and representation (i.e type) match an element in
the list.

if the type or contents don't match with the pre-defined list, false is
returned else true.