Linter: Empty case for if/ConditionalAssignment
This commit fixes two linter errors that have been raised during ci
build.

The first is switching empty case statements for if and the other is
rubcop recommending the use of return of the conditional for variable
assignment and comparison.

These have not affected any regression test artefacts and flow