Tidy up Post Code lookup error handling
When a postcode cannot be found in Imminence API, the endpoint
returns a `404` status. We should accept this error as standard control
logic, and raise a corresponding error within our own application. This
shows the user that the problem is to do with their input.

Any other error is an unexpected error, and should raise an exception
showing that the problem with lookup is at our end.

This change also removes the special logic for `GdsApi::BaseError`. We
should never get this error, and if we do, it’s truly exceptional and
the website should fail “noisily”. Removing this `rescue` block will
log the error directly to Errbit, and show the user a generic `500`
page, whilst also tidying up our own control logic.