Preparatory: Supply controller to erb renderer
All erb templates are rendered by the custom [1] smart answer renderer

This [1] ActionView instance only setups the view_paths context and not
the controller.

This makes it difficult to learn what the request context is, here it is
being included as part of the arguments of the presenters and
eventually passed to the ErbRenderer [1]

This was not added initially, to avoid  the precedence given to controller
prefixes which override the search for templates.

When the prefixes are set, templates are looked up at the controller paths
eg: ['smart answers', 'application']. After the controller is added the
prefixes are set to an empty array, to allow the template and registry load
paths to be used. See [2] and [3] for more details.

[1] https://github.com/alphagov/smart-answers/blob/3e784b05ad09d91af200b4dacb475f56777956a0/lib/smart_answer/erb_renderer.rb#L17
[2] https://github.com/rails/rails/blob/master/actionview/lib/action_view/base.rb#L194
[3] https://github.com/rails/rails/blob/master/actionview/lib/action_view/lookup_context.rb