Wrap each partial with debugging div
When a content designer need to edit smart answer content, they
find it difficult to find the relevant templates and partials
that a particular outcome or question uses.

This commit coninues the spike from [1] with the aim to wrap
and tag each partial for later identification, highlighting
and linking to the relevant github partial template.

An example can be found when looking at templates like [2],
that have nested partial templates within a content_for title.

[1] github.com/alphagov/smart-answers/pull/2642
[2] /marriage-abroad/y/sweden/uk/partner_british/same_sex