Accommodate title with debug div wrapper
Following the previous commit where the debug div wraps each partial
template.

As a result, content_for title produced a wrapping h2 tag that wrapped
the debug div. This breaks the presentation of any title with a partial

Here the solution has been to extract the title and path and present
them with matching debug div wrap where required.