Allow viewing Govspeak on all environments
This commit removes the EXPOSE_GOVSPEAK constant and in effect allowing
users to view smart answer Govspeak on all environments.

This has become necessary as sometimes the staging and production get
provisioned and the EXPOSE_GOVSPEAK appears unset making it hard for
content designers to view Govspeak equivalent of the html page.