Display an alert on heroku deployments
We want to avoid the general public being confused about the
information in the Smart Answers preview apps on Heroku.

We are then displaying a big alert banner on top of each page of heroku
deployments, to make very clear that the information shouldn't be
relied upon, and the user should instead refer to gov.uk for the
official information.