Add payment questions to maternity flow
The maternity paternity calculator sometimes recommends some
overpayments for some users.

To fix this, extra questions are required to help the tool produce
more accurate calculation and recommendations.

This change maybe required across the maternity, paternity and adoption
flow of this calculator.

This commit introduces one question each for weekly and monthly payments
but only affecting the maternity flow.

The regression tests artefacts have been updated and mostly have been
renamed as a result of the new questions.