Preparatory: Add new payment options
Each payment option question that has been added to the
maternity, paternity and adoption flows and only contain
option 2 and 8 that have already been defined.

This commit adds the 9 and 10 payment options for weekly
and 3 for monthly respectively.

The aim is to add these new payment option in preparation
for changes to the average weekly earnings.

The regression test artefacts generated have only been created
with current content and calculations.