Adoption: Move average_weekly_earnings block
This commit moves the average_weekly_earnings and above_lower_earning_limit
blocks to the outcome block.

This has been done because the introduction of payment_option gets set
after the earnings_for_pay_period_adoption question.

This brings to effect the changes to the average_weekly_earnings to the
regression test artefacts.