Re-define arrays of countries
In order to reduce the redundancy of the code required to test
marriage abroad outcomes for each country, we are going to define a single
set of instructions to test the outcomes.

Before doing this we are defining the lists of countries:
- with 18 outcomes
- with 6 outcomes
- with 2 outcomes
- countries not yet flattened

Each list is included in an array with the corresponding name.