Use DateRange for Agricultural Holiday Entitlement
We are using DateRange to simplify the code which involves
calculation with date ranges.

`AgriculturalHolidayEntitlementCalculator.available_days` calculates
the number of days between the 2 dates, without including the first
day in the date range.
For this reason we added the method `non_inclusive_days`,
beside `number_of_days`, inside the `class DateRange` to make this
calculation.