Fix minimum wage
Calculate the basic_rate as the average between the normal rate and
the overtime_hourly_rate, instead of just returning the overtime_hourly_rate.

This is stated here: https://www.gov.uk/overtime-your-rights/overview.

This prevents the message "You are not getting the minimum wage" to appear, if
the basic rate is above the minimum rate.

This includes also:
- Update minimum_wage_calculator_test values on the basis of the new
hourly rate average calculation
- Update checksums and artefacts