Add combined question template
This new question template is a combination of
`did_the_employee_work_for_you?` and `is_the_employee_on_your_payroll?`.

Instead of asking separately:
"Did the employee work for you on or before 1 April 2017?"

and

"Was the employee (or will they be) on your payroll on 17 September
2017?"

We will instead ask:

"Did the employee have a contract with you between 1 April 2017 and
17 September 2017?"

Subsequent commits will incorporate this new template into the flow.