Display values less than 1 in pence not pounds in format_money
We do this by scaling the value by 100 when it's less than 1 (to get it in
pence scale), and then passing it to the `number_to_currency` helper with
a `p` unit and `<value><unit>` format (instead of the default
`<unit><value>` format).  This means we get 10p instead of £0.10, but also
that we get 5p instead of 05p.  We handle negative amounts correctly and
scale back to pounds after -£1.  We make sure to leave true 0 as £0 though
and not 0p.

In passing we also make sure that any value that would be rendered as a
whole number of pounds (e.g. £1.00001, and £1.00) is rendered without
decimals.

To be able to to this we need to extract a numeric value from the provided
amount so we can check if it is between 1 and -1, and check if it (or the
rounded to 2 places version of it) is zero.  If we can't extract a number
from the value then we proceed as we used to and pass the amount to the
number_to_currency method with the assumption the amount is in pounds.