Improve arbitary input handling of CurrentQuestionHelper#default_for_date
We recently received alerts in [Sentry][1] for 500 errors when receiving certain inputs for a date. These alerts arose from passing an array response to a date question. Our handling of arbitrary inputs to the `CurrentQuestionHelper#default_for_date` was no robust enough to deal with this, and attempted to call `.to_i` on an `Array`, which throws a `NoMethodError`.

This change makes the method more robust by direclty attempting to cast to `Integer`, and rescuing errors by returning `nil`. This change also adds a number of tests to cover various types of input to the  method.

[1]: https://sentry.io/govuk/app-smart-answers/issues/396675536/