Switch to taxonomy navigation for tagged content
When a content item is tagged to the worldwide taxonomy or
has taxon links and isn't tagged to mainstream browse,
we display the new taxonomy navigation.
This commit renames the EducationNavigationABTest concern
to TaxonomyNavigation and alters the logic according to the
conditions above. Also remove some of the AB test header and
beta banner code.