Fix Maternity Smart Answer Error
There's a problem with the calculator when the last SMP payment is after the last working day
of the month. This was spotted for June 2018, but we think it would happen whenever this
scenario arises.

Previously the code iterated the SMP payments according to the last working day of each month
but it did not tackle the scenario where the final SMP payment is outside of the last working
day of leave.

The new code iterates through the last working day of each month while the last working day of
the month is less than the final SMP pay date. This ensures that all payments are made. In this
scenario an extra month is added to which the last payment will be made.