Fix state pension calculator for leap year births
This issue affects users born after 5th April 1976 on a leap year . These users will currently be eligible for their state pension on their 68 birthday (which will also be a leap year).

Currently users are given the correct entitlement date for their state pension but says their state pension age will be 67 years, 11 months, 30 days. This is because the calculator deducts a day for leap year births, but does not take into account the state pension date also being a leap year.

This fix updates the calculator to check whether the state pension date is also a leap year. If so. no adjustment needs to be made to the state pension age.