Use standard buildProject call in Jenkinsfile
Our job in jenkins is 'smartanswers', but our repo on github is
'smart-answers'.  The default buildProject used to not understand this
difference which is why we didn't use it.  However, it's now been updated
so that it does understand it and we no longer need our custom version.

We also need to:

1. configure an extra build param: RUN_REGRESSION_TESTS - used to control
   if we should run the (slow) regression tests or not
2. customise the test task so that it runs the default tests, but also
   will run the regression tests if we're on the master branch, or the
   RUN_REGRESSION_TESTS param is set.