Abstract TaxYear class to a year range with a fixed start day and month
We have a `TaxYear` class which represents a year range which always
starts on the 6th April. We have other instances of this sort of
behaviour (for example Agricultural Holiday years, which start on
1st October).

This change abstracts the Tax Year logic into a higher level class,
which will work with arbitrary year start dates.