Add a question asking if the adoption is from abroad
This change adds a question to the Adoption calculator, establishing
if the child is being adopted from the UK or from abroad.

The answer to this question will be used to modify the copy of some of
the later questions, and to add validation for the adoption leave start
date, which is different depending on if the child is being adopted from
abroad.