Improve DateRange negative case handling
If the end date is initialised to a date before the begins date,
`DateRange` silently sets it to the day before the begins date. This was
probably done to make it return 0 number of days. This is a strange
behaviour and could lead to unexpected results.

Instead of automatically set the end date, this PR allows the end date to
be before the begins date and makes DateRange return the exact amount of
negative number of days. The `empty?` method is also set to return
true in a case of negative or zero number of days.

This avoids to break the other related calculations (e.g.
`gap_between`) and doesn’t make changes to the end date initially set.