Update Regression test expected outputs
When upgrading to Ruby 2.4.2 - there were changes made in how rounding
happens to certain negative numbers

Under Ruby 2.3.1 - -35.355.round(2) would return -35.35, whereas under
Ruby 2.4.2 it returns -35.36. As the latter behaviour is more consistent
with other rounding, we have considered it to be the correct case, hence
the updates to the expected outputs

E.g. Ruby 2.3.1:
- 2.355.round(2) = - 2.36
-35.355.round(2) = -35.35
-70.355.round(2) = -70.36

Ruby 2.4.2
- 2.355.round(2) = - 2.36
-35.355.round(2) = -35.36
-70.355.round(2) = -70.36