Stop stripping organisation info from content item
We stripped organisation info from the content item presented to the
view because it would result in the govuk:analytics:organisations meta
tag being set and this would be sent to Google Analytics (GA) as a custom
dimension. We did this because the tag was mostly used by whitehall and
departmental users wouldn't expect to see mainstream content appearing
in their GA reports for their department.

During migration the mainstream formats ported to government-frontend
started sending this info to GA and it didn't cause any problems. This
means we're happy to stop stripping this data and have all content items
that have organisation info send it to GA.