Move flow calculations into Calculator for student finance calculator
Refactor maintenance loan and maintenance grant logic by moving it out
of the flow class and into the calculator class. This helps keep the
flow class focused on routing, as described in doc/refactoring.md

This will make it easier to use the result of this calculation in other
outcomes.