Ensure 'Before you start' subheader is rendered on start page
Add the missing header to make the template consistent with the one in
frontend, which is normally used to render start pages
(https://github.com/alphagov/frontend/blob/master/app/views/simple_smart_answers/show.html.erb).

This fixes the header when previewing the app, either in development or
on a branch deployment.