Fix the "Start now" button appearance in preview
Update the div and link classes to be consistent with the equivalent
template in the frontend app
(https://github.com/alphagov/frontend/blob/master/app/views/simple_smart_answers/show.html.erb).

Frontend is normally used to render smart answer start pages. The
template updated here is used to preview the page, either in dev or on
branch deployments.

This change adds the missing '>' arrow to the "Start now" button.