Introduce `assert_same_url` helper for integration tests
This is an abstraction of the `assert_current_url` helper that already
exists that allows you to pass in both urls you want to compare instead
of only checking against the current page url.  We then use this in two
integration tests that compare the action of a form element.  Upgrading
poltergeist broke these tests because the action of the form element was
no longer reported as a path, but a full url.  Using assert_same_url fixes
this by comparing just the paths (and optionally the query strings).