Remove breadcrumbs from all pages
The breadcrumbs were introduced in
https://github.com/alphagov/smart-answers/pull/2805 (by me).

At this time this app also rendered the start pages like
https://www.gov.uk/calculate-agricultural-holiday-entitlement. Those
have since moved to government-frontend.

This PR removes the breadcrumbs from the pages because they don't add
any information for the user. Smart answer pages are kind of
transaction-y pages that don't need navigation elements because people
need to focus on answering the questions.

https://trello.com/c/4ffi5zAj