Remove out of date options from minimum wage flow
You can only check 6 years into the past, so presenting choices for going
further back than that is not useful.  This also means we can remove an
entire outcome: "does_not_apply_to_historical_apprentices".  This outcome
was for apprentices who started before 2010 because there was no minimum
wage for apprentices before that.  Now that we don't allow checking before
2012 there is no need for this outcome.