Add a way to configure a 'none' checkbox option.
In order to make it easier for people to explicitly
select none of the given options, add a way to configure
a 'none' option to a checkbox question. The label and
option prefix are configurable.
Previous no-input functionality remains untouched if
the configuration method 'set_none_option' is not called
within the checkbox_question block.