Remove marriage abroad CI check
This is causing CI builds to time out, and the instruction is listed in the docs for marriage abroad, so hopefully this won't be too much of an issue now.

If it is then we should fix this more broadly, probably by having jenkins (reliably) do all this automation for us.