Fixes tests and makes variables more dynamic
This commit changes the way variables are passed through each context block. We’ve done this because they were previously quite brittle and had a number of hard-coded values that would have to be changed every leap year. Now the number of days in a year is dynamically calculated by the number_of_days method of the YearRange class.

Co-authored-by: Rosa Fox <rosa.fox@digital.cabinet-office.gov.uk>