Implement new EEA travel rules
As per published guidance, these are the rules for travel to the UK from an EEA country if the UK is outside the EU at the time.

https://www.gov.uk/guidance/european-temporary-leave-to-remain-in-the-uk#who-will-need-to-apply

All EEA citizens can stay in the UK for up to and including 3 months without a visa.  Beyond that they will need a European Temporary Leave to Remain.

Republic of Ireland citizens do not need a visa even if staying in the UK longer than 3 months.