Don't round average weekly earnings
We should be using the full 7 decimal place value throughout the
calculation and only round it at the end, rather than rounding it during
the calculation.

I've taken the code from
https://github.com/alphagov/smart-answers/blob/cddae6f5c91c918c7164345e06679121d04b6c8e/lib/smart_answer/calculators/maternity_pay_calculator.rb#L128
and rather than calling the method I've made it an explicit call so it's
very clear where the code has come from.