Ignore definite article "the" from country names when alphabetising drop downs
The Worldwide Locations API returns a specific order of country names to display those
starting with `The` in the most convenient position for the user, for example
`The Gambia` appears under `G` rather than `T`. However, the CountrySelect may add some additional
countries and then re-order the list.

This PR ensures that SmartAnswers continues to display the user-friendly order by ignoring the definite
article `The` at the beginning of any country names when sorting.