Guard against incorrect contractual_days_covered_by_earnings values
In the statutory sick pay calculator. Previously with Ruby 2.5,
`BigDecimal('4 weeks')` would be 4, but with Ruby 2.6, it raises an
exception.

With the previous behaviour, users were getting incorrect answers, as
the calculator was interpreting '4 weeks' as 4 days, so add validation
to prompt the user to enter a number of days.