This is to give a more accurate outcome.
From Content team:

An earlier version of the smart answer assumed you weren't eligible for
Statutory Maternity Leave if you weren't eligible for Statutory
Maternity Pay - that's true in 99% of cases but not all.

We fudged the outcome by adding 'The employee is entitled to Statutory
Maternity Leave if... they have an employment contract with you,
regardless of when it started', but it's better to have proper routing
and make the outcomes easier to understand.

Trello card: https://trello.com/c/QUboSszZ/1159-employers-maternity-calculator-new-question-in-maternity-branch