Updated application process
People can no longer get an affidavit or affirmation to prove they can marry in Belgium, or make an appointment. They must now get a ‘certificate of custom and law for marriage' and apply by post. All documents someone needs to send with their application are now listed on the application form, including if someone is divorced or widowed.